requirements.txt
anyio==4.12.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.14.3
beniget==0.4.1
bleach==6.3.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.19
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dotenv==0.9.9
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fasteners==0.14.1
fastjsonschema==2.21.2
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.5.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
ipykernel==7.1.0
ipython==8.37.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.2.1
mistune==3.1.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.10.3
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.5.3
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.5.1
ply==3.11
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.12.4
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-json-logger==4.0.0
python-telegram-bot==22.6
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
reportlab==3.6.8
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.8.1
stack-data==0.6.3
sympy==1.9
systemd-python==234
terminado==0.18.1
testpath==0.5.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
